+++ web/src/message_edit.ts
@@ -248,8 +248,8 @@ export function is_stream_editable(message: Message, edit_limit_seconds_buffer =
     }
 
     // Organization admins and moderators can edit stream indefinitely,
-    // irrespective of the stream editing deadline, if
-    // move_messages_between_streams_policy allows them to do so.
+    // irrespective of the stream editing deadline, if they are in the
+    // can_move_messages_between_channels_group.
     if (current_user.is_admin || current_user.is_moderator) {
         return true;
     }