+++ zerver/lib/event_schema.py
@@ -1067,6 +1067,7 @@ group_setting_update_data_type = DictType(
         ("can_delete_any_message_group", group_setting_type),
         ("can_delete_own_message_group", group_setting_type),
         ("can_manage_all_groups", group_setting_type),
+        ("can_move_messages_between_channels_group", group_setting_type),
         ("direct_message_initiator_group", group_setting_type),
         ("direct_message_permission_group", group_setting_type),
     ],